The replacement of 'int' with 'size_t' caused a number of errors across multiple platforms in our CI system.  The culprit is not obvious, so I am reverting the commit until we can take a closer look at the underlying issue.
Revert "1. Attempted to reduce the number of conversion errors"

This reverts commit 1b941d342d5614aec5f55f819c0ee590e8bd3d6e.

+++ include/ncdispatch.h
@@ -358,7 +358,7 @@ extern int NCDAP_ping(const char*);
 
 /* Misc */
 
-extern int NC_getshape(int ncid, int varid, size_t ndims, size_t* shape);
+extern int NC_getshape(int ncid, int varid, int ndims, size_t* shape);
 extern int NC_is_recvar(int ncid, int varid, size_t* nrecs);
 
 #define nullstring(s) (s==NULL?"(null)":s)
